Lab 2 finished v4, finished all DebugExercises.

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,8 +5,10 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) {
+        float quotient = top / bottom;
+//        float quotient = (float) top / (float) bottom;
+//        System.out.println(quotient);
         int result = Math.round(quotient);
         return result;
     }
